v9.4: аудит + читы + A7800-fix + GBA (gpSP) вшит + integer scale
- A7800: static a7_ram[16K] + a7800_set_memory() — чинен краш memory_ram=NULL
- читы: поиск по базовому имени без (...)[], BOM/CRLF, буфер 32K, CHEATS_MAX 128
- читы: ручной ввод геймпадом (Game Genie: слоты + палитра символов)
- читы: протечка между играми закрыта (cheats_reset при обычном Enter)
- читы: UART-мусор убран (диагностические printf удалены)
- меню/браузер: dirty-флаг рендера — мерцание устранено
- джойстик: edge-detect + замедленный автоповтор (400ms/200ms)
- вход в подменю: ожидание отпускания кнопок (не активирует первый пункт)
- NES FCEUmm: FKB fix, диспетчер run_emulator, чит-меню ESC=отмена
- GBA host (gpSP): init/run_frame/render, ROM из памяти (g_ram_rom),
  BIOS open-source 16K, GBA_RENAME (objcopy: vram/reg/cheats/init_memory/
  init_cpu/load_bios -> gpsp_* против конфликтов с FCEUmm/GPGX)
- integer scale: портативные (GB/GBA/Lynx/NGP/GG) в родном разрешении x3-x5
- emu_scale: инкрементальный без делений (~5-10 мс/кадр экономии)
- OHCI: per-port ED/TD пул (g_head_ed[2], g_td[2][8])
- SD/FAT/UART/USB warnings нашего кода: 0

Known issues (доработка в следующих сессиях):
- GBA: игры с EEPROM/SRAM виснут после логотипов (не мапится gamepak_backup)
- GBA: NGP/RACE белый растр на части ROM (та же нехватка write-map backup)
- Neo Geo MVS (roms/neogeo) — не эмулируется (нужно отдельное ядро FBA/MAME)

+++ h3_bare/platform/fb/h3_hdmi.c
@@ -314,7 +314,10 @@ static void hdmi_pll_set(uint32_t clk_khz) {
 
 	clock_set_pll_video_factors(best_m, best_n);
 
-	printf("dotclock: %u kHz = %u kHz: (24MHz * %u) / %u / %u\n", clk_khz, (uint32_t)(clock_get_pll_video() / 1000) / div, (uint32_t)best_n, (uint32_t)best_m, (uint32_t)div);
+	printf("dotclock: %lu kHz = %lu kHz: (24MHz * %lu) / %lu / %lu\n",
+	       (unsigned long)clk_khz,
+	       (unsigned long)((uint32_t)(clock_get_pll_video() / 1000) / div),
+	       (unsigned long)best_n, (unsigned long)best_m, (unsigned long)div);
 }
 
 /*
